nowcoder/requirements.txt
requests>=2.25.0
python-dotenv>=0.15.0
beautifulsoup4>=4.9.0
playwright>=1.40.0
openai>=1.0.0
